fix user settings form
